requirements.txt
opencv-python-headless==4.8.1.78
mediapipe==0.10.9
reportlab==4.1.0
anthropic>=0.88.0
Pillow==10.2.0
numpy==1.26.4
streamlit>=1.50.0
